fix: IoT.

+++ SmartCollar/Application/Schema/smart_collar_schema.py
@@ -7,7 +7,7 @@ from appointments.domain.models.availability import Availability
 class SmartCollarRequest(BaseModel):
     serial_number: str
     temperature: float
-    lpm: int
+    lpm: float
     battery: float
     location: LocationType
 
@@ -17,7 +17,7 @@ class SmartCollarRequest(BaseModel):
 
 class SmartCollarUpdateRequest(BaseModel):
     temperature: float
-    lpm: int
+    lpm: float
     battery: float
     location: LocationType
 
@@ -29,7 +29,7 @@ class SmartCollarResponse(BaseModel):
     id: int
     serial_number: str
     temperature: float
-    lpm: int
+    lpm: float
     battery: float
     location: LocationType
     pet_id: Optional[int]  # Allowing pet_id to be None
