+++ SmartCollar/Domain/Models/smart_colllar_model.py
@@ -13,7 +13,7 @@ class SmartCollar(Base):
     id = Column(Integer, primary_key=True, index=True)
     serial_number = Column(String(255), unique=True, nullable=False)
     temperature = Column(Float, nullable=True)
-    lpm = Column(Integer, nullable=True)
+    lpm = Column(Float, nullable=True)
     battery = Column(Float, nullable=False, default=100.0)
     latitude = Column(Float, nullable=True)
     longitude = Column(Float, nullable=True)
